remove kernel module
