+++ README.md
@@ -41,33 +41,24 @@ See [docs](./docs/.)
  * `src/binary/` - binary representation of data to store/read it on/from disk
  * `src/binder/` - module responsible for executing `Bind` message from client from extended query PostgreSQL protocol
  * `src/constraints/` - mechanics to ensure data validity that are going to store
+ * `src/connection/` - mechanics to establish connection between clients and server
  * `src/data_manager/` - module that manages how to, where store/read data and metadata
  * `src/description/` - building blocks of statement description, the info per statement is used to send statement description
                         on client request of `Describe` message
  * `src/expr_eval/` - module that evaluates static and dynamic expression in transformed `ast`
- * `src/kernel/` - core concept of the system. All modules (except `protocol`) depends on it.
-                   It should provide conceptual abstraction for other modules. Good examples
-                   are `SystemResult` and `SystemError`. Other part of system uses them to
-                   handle errors that are not local for a module but more system wide.
  * `src/meta_def/` - database structure definitions (table, columns etc)
- * `src/metada/` - module responsible for managing on disk data about data
+ * `src/metadata/` - module responsible for managing on disk data about data
  * `src/node/` - database node (member, server, instance) code. Handles network communication
                  with clients and process management of incoming queries. It also contains
                  concrete `trait` implementations from `src/protocol/` module.
  * `src/parser/` - wrapper around `sqlparaser-rs` crate
+ * `src/pg_model/` - intermediate code between [pg_wire](https://github.com/alex-dukhno/pg_wire) and other
+                     database management system parts. Most probably code will be moved to `pg_wire` in incremental steps.
  * `src/plan/` - building blocks for execution query plans
- * `src/protocol/` - server-side (backend) API of 
-                    [PostgreSQL Wire Protocol](https://www.postgresql.org/docs/12/protocol.html)
-                    The goal is to provide high level `trait`s and `struct`s to help other `rust`
-                    database implementations be `PostgreSQL` compatible.
-                    **Must** not depend on any modules of the system and has as small as possible
-                    dependencies on other crates because it is intended to move out of the project
-                    into completely separate crate. Right now it is in the project because of ease
-                    of testing, prototyping and development.
+ * `src/repr/` - building blocks to how transform binary representation into values
  * `src/query_analyzer/` - module responsible for query analysis/description
  * `src/query_exeuctor/` - module to execute incoming `SQL` queries. That is it.
  * `src/query_planner/` - module uses `sqlparser::ast` to generate a plan how to execute query
- * `src/repr/` - building blocks to how transform binary representation into values
  * `src/sql_model/` - module contains code to support different `SQL` types other `SQL`
  * `src/storage/` - module abstracts work with `schema`s, `table`s and other on disk structures
  * `tests/compatibility` - groovy based tests to check compatibility with [PostgreSQL](https://www.postgresql.org/)
