+++ src/data_manager/Cargo.toml
@@ -7,7 +7,6 @@ publish = false
 
 [dependencies]
 binary = { path = "../binary" }
-kernel = { path = "../kernel" }
 meta_def = { path = "../meta_def" }
 metadata = { path = "../metadata" }
 repr = { path = "../repr" }
