--- src/kernel/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "kernel"
-version = "0.1.0"
-authors = ["Alex Dukhno <alex.dukhno@icloud.com>"]
-edition = "2018"
-publish = false
-
-[dependencies]
-backtrace = "0.3.50"
